Get the harmony test suite building.
Support_Jetty needs to be rewritten in terms of MockWebServer. I've removed the
concurrent tests on the assumption that we'll keep those separate for now (and
that they won't be updated by harmony ever again, even though the true upstream
will).

I've removed all the 'main' methods (which add a dependency on code we don't
have), and I've added the harmony support tree, which was classlib/support
rather than classlib/modules/support, so didn't get included last time.

Change-Id: Ifba2d420bb1cc4c3e4ee76d6a8ee68e7cf846ed9