ThreadReference.OwnedMonitorsStackDepthInfo is now implemented.
Note that the test was hard-coding stack depths that depend on the
library implementation. Since the stack frames in question are the
outermost two (with all the library frames after them), I've made
the smallest change of asking for the total number of stack frames
and then asserting that we're dealing with the outermost two.
Getting the actual stack frames and checking the actual methods
would be a lot hairier, and not obviously valuable (since we already
test those JDWP requests elsewhere).

Change-Id: Ie56bc862639338d6ae02bc1c08fad3b1e7ed4ef1