Fix OwnedMonitorsStackDepthInfoTest.
We were still relying too much on implementation details, in particular
"where does the test thread actually get suspended?" --- it turned out
that we were usually running fast enough to suspend at the receive, but
sometimes we could suspend at the preceding send.

Also improve the test code so that we check multiple stack frames in
the test code, both with and without held monitors, and from synchronized
methods as well as from synchronized blocks.

Change-Id: Ibda449d097d99b15121076b7cad75383b76ed35d