Fix reply check in SingleStepTest.
Tests in SingleStepTest class all use the "stepFunction" method. However this
method badly checks we succeeded in clearing the single-step event. Therefore
these tests could all pass even if clearing the single-step event fails.

This CL fixes it by checking the reply corresponding to the EventRequest.Clear
command, not the reply of a previous command.

Change-Id: I9256b3da4a1ac905a5ed29bf792d5c3ab1f97267