Fix OwnedMonitorsStackDepthInfo JDWP tests
Ensure we signal the debuggee to continue at the end of the test. It
should prevent the test from failing because the debuggee remains
blocked while we wait for it to finish.

We set the continue message in the 'finalSyncMessage' field. When the
test ends, it will automatically send the message in this field to
the debuggee.

Bug: 17751850
Change-Id: I3be530230c4ad1b768be18b2136f5580a0af4a6d