Fix OwnedMonitorsStackDepthInfo JDWP tests
Ensure we signal the debuggee to continue at the end of the test. It
should prevent the test from failing because the debuggee remains
blocked while we wait for it to finish.

We set the continue message in the 'finalSyncMessage' field. When the
test ends, it will automatically send the message in this field to
the debuggee.

Bug: 17751850
(cherry picked from commit 1c40fe5dbb39eaa6eed3fde3f06effe3517b32d0)

Change-Id: I9b0960871db90b035491fe5c061ab7c9459ddc4c