Test event requests with LocationOnly modifier other than breakpoint
Checks we correctly support LocationOnly modifier for all possible
events other than breakpoint: single-step, exception, field access
and field modification events.

For exception and field access/modification events, we throw or
access/modify a field in two different methods. Then we request
corresponding event with a LocationOnly modifier related to the
second method only.
Since a location is {class id, method id, code index}, we'd need to
know exactly the code index where the event can be reported. As we
do not know the code (instructions and their respective code index),
we simply send an event request for each location. The common
testing code is in EventLocationEventTestCase class.

For single-step, we can't do the same because we can't request
multiple single-step events on the same thread. So we suspend in a
method, then single-step to the location of the last line of this
method.

Also move "resume debuggee" message into JDWPTestCase.resumeDebuggee
method so this is printed for all tests using it.

Bug: 17908144
(cherry picked from commit 684d91ef9b8c8022f19cc0d51598c2402303ed7c)

Change-Id: Iaa9698f6b16239a779f3afe7870fc66c9bcf553b