Fix flaky test ThreadStartTest
We may receive THREAD_START events for runtime internal threads (like
GCDaemon) before receiving one for the tested thread.

We now loop until receiving a THREAD_START event for the tested
thread. If we receive a THREAD_START event for another thread, we
just ignore it and wait for the next event.

Bug: 19438651
Change-Id: I18f61c48e08c77817a4e75d2c277a101cc9b9f0e