Use 127.0.0.1 to avoid DNS lookups.
In the way we setup the chrome buildbot testing, DNS lookup
does not work: ART is being pushed like an app but brings
its own version of bionic, which can be incompatible with
the version on the device. DNS lookup talks to a daemon, which
understands the old version of bionic, not the new one.

Change-Id: I761a385b12b6c4b385f1f2c7154b4776adb7d869